customAppDemo/requirements.txt
django
djangorestframework
djangorestframework-gis
pyyaml
requests
django-cors-headers
psycopg2-binary
gdal-bin
